+++ template/src/hooks/useColor/index.js
@@ -1,0 +1,1 @@
+export {default} from './useColor'
