+++ template/src/index.js
@@ -3,24 +3,18 @@ import 'regenerator-runtime/runtime'
 import {registerHtml, start} from 'tram-one'
 import ColorHeader from './components/ColorHeader'
 import './styles.css'
-import { useColor } from './hooks/ColorHook'
+import useColor from './hooks/useColor'
 
 const html = registerHtml({
 	ColorHeader
 })
 
 const home = () => {
-  const [color, incrementColor] = useColor()
+  const [color] = useColor()
 	return html`
     <div>
       <ColorHeader />
-      Thank you for using Tram-One!
-      <div
-        style="color:${color}"
-        onclick=${incrementColor}
-      >
-       The color is controlled by global state 
-      </div>
+      <div style="color:${color}"> Thank you for using Tram-One! </div>
     </div>
   `
 }
